Move watcher code to watch.go
