+++ internal/pkg/service/common/etcdop/prefix_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"go.etcd.io/etcd/api/v3/mvccpb"
 
 	"github.com/keboola/keboola-as-code/internal/pkg/service/common/etcdop/serde"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils/etcdhelper"
@@ -391,41 +390,6 @@ func BenchmarkPrefixT_DeleteAll(b *testing.B) {
 			b.Fatalf("unexpected result")
 		}
 	}
-}
-
-func TestTypedPrefix_Watch(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	etcd := etcdhelper.ClientForTest(t)
-
-	pfx := typedPrefixForTest()
-	ch := pfx.Watch(ctx, etcd, func(err error) {
-		assert.FailNow(t, err.Error())
-	})
-	go func() {
-		key1 := pfx.Key("key1")
-		err := key1.Put("foo").Do(ctx, etcd)
-		assert.NoError(t, err)
-	}()
-	msg := <-ch
-	fooVal := fooType("foo")
-	assert.Equal(t, EventT[fooType]{
-		Type:  mvccpb.PUT,
-		Key:   "my/prefix/key1",
-		Value: &fooVal,
-	}, msg)
-
-	go func() {
-		key1 := pfx.Key("key1")
-		_, err := key1.Delete().Do(ctx, etcd)
-		assert.NoError(t, err)
-	}()
-	msg = <-ch
-	assert.Equal(t, EventT[fooType]{
-		Type: mvccpb.DELETE,
-		Key:  "my/prefix/key1",
-	}, msg)
 }
 
 func typedPrefixForTest() PrefixT[fooType] {
